moved stats from BACK to FRONT
